Add files via upload
My apprenticeship at GM ends the last day of June and started practicing algorithms again to prepare for interviews.
The two sum algorithm is a good one for interviewers to ask, so they can quickly gain an understanding of a candidates depth of knowledge.  To solve this a naive N squared brute force approach can be used, but what they are really looking for is something faster using a hashmap.